requirements.txt
seaborn
matplotlib
pandas-profiling==3.6.5